+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [JENKINS-55345](https://issues.jenkins-ci.org/browse/JENKINS-55345): 
 DocFX Parser: Ignore Info messages and do not treat them as warnings.
+- [JENKINS-42823](https://issues.jenkins-ci.org/browse/JENKINS-42823): 
+msbuild Parser: Fix false positive with parser.
 
 ## [3.0.0](https://github.com/jenkinsci/analysis-model/compare/analysis-model-2.1.2...analysis-model-3.0.0) - 2019-3-15